agent_app_sample_code/pyproject.toml
[tool.poetry]
name = "genai-cookbook-utils"
version = "0.1.0"
description = ""
authors = ["Eric Peter <epeter@gmail.com>"]
readme = "README.md"
packages = [{include = "utils"}]

[tool.poetry.dependencies]
python = "^3.11"
databricks-connect = "~15.2"
pydantic = "^2.9.2"
pyyaml = "^6.0.2"
databricks-vectorsearch = "^0.42"
databricks-sdk = {extras = ["openai"], version = "^0.36.0"}
mlflow = "^2.17.2"
databricks-agents = "^0.9.0"
pymupdf4llm = "0.0.5"
pymupdf = "1.24.13"
markdownify = "0.12.1"
transformers = "4.41.1"
torch = "2.3.0"
tiktoken = "0.7.0"
langchain-text-splitters = "0.2.0"
ipykernel = "^6.29.5"
hatchling = "^1.25.0"
pypandoc-binary = "1.13"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
